Put special split_last behaviour in private module and re-use in tree and file_system
